+++ libff/algebra/field_utils/field_utils.hpp
@@ -9,11 +9,90 @@
 #define FIELD_UTILS_HPP_
 #include <cstdint>
 
-#include <libff/algebra/field_utils/bigint.hpp>
-#include <libff/common/double.hpp>
-#include <libff/common/utils.hpp>
+#include "libff/algebra/field_utils/bigint.hpp"
+#include "libff/common/double.hpp"
+#include "libff/common/utils.hpp"
+
+#include "libff/algebra/fields/binary/gf64.hpp"
+#include "libff/algebra/fields/binary/gf128.hpp"
+#include "libff/algebra/fields/binary/gf192.hpp"
+#include "libff/algebra/fields/binary/gf256.hpp"
+#include "libff/algebra/fields/prime_base/fp.hpp"
 
 namespace libff {
+
+template<typename FieldT>
+struct is_additive {
+    static const bool value = false;
+};
+
+template<>
+struct is_additive<gf64> {
+    static const bool value = true;
+};
+
+template<>
+struct is_additive<gf128> {
+    static const bool value = true;
+};
+
+template<>
+struct is_additive<gf192> {
+    static const bool value = true;
+};
+
+template<>
+struct is_additive<gf256> {
+    static const bool value = true;
+};
+
+template<typename FieldT>
+struct is_multiplicative {
+    static const bool value = false;
+};
+
+template<mp_size_t n, const bigint<n>& modulus>
+struct is_multiplicative<Fp_model<n, modulus>> {
+    static const bool value = true;
+};
+
+enum field_type {
+    multiplicative_field_type = 1,
+    additive_field_type = 2
+};
+
+template<typename FieldT>
+field_type get_field_type(const typename enable_if<is_multiplicative<FieldT>::value, FieldT>::type elem);
+
+template<typename FieldT>
+field_type get_field_type(const typename enable_if<is_additive<FieldT>::value, FieldT>::type elem);
+
+template<typename FieldT>
+std::size_t log_of_field_size_helper(
+    typename enable_if<is_multiplicative<FieldT>::value, FieldT>::type field_elem);
+
+template<typename FieldT>
+std::size_t log_of_field_size_helper(
+    typename enable_if<is_additive<FieldT>::value, FieldT>::type field_elem);
+
+template<typename FieldT>
+std::size_t soundness_log_of_field_size_helper(
+    typename enable_if<is_multiplicative<FieldT>::value, FieldT>::type field_elem);
+
+template<typename FieldT>
+std::size_t soundness_log_of_field_size_helper(
+    typename enable_if<is_additive<FieldT>::value, FieldT>::type field_elem);
+
+template<typename FieldT>
+std::size_t get_word_of_field_elem(
+    typename enable_if<is_additive<FieldT>::value, FieldT>::type field_elem, size_t word);
+
+template<typename FieldT>
+std::size_t get_word_of_field_elem(
+    typename enable_if<is_multiplicative<FieldT>::value, FieldT>::type field_elem, size_t word);
+
+template<typename FieldT>
+FieldT coset_shift();
 
 // returns root of unity of order n (for n a power of 2), if one exists
 template<typename FieldT>